requirements.txt
tensorflow
streamlit
opencv-python
numpy
pillow
matplotlib
